+++ cognite/neat/rules/models/entities/__init__.py
@@ -1,4 +1,7 @@
-from ._base import (
+from ._constants import EntityTypes, Undefined, Unknown
+from ._loaders import load_connection, load_dms_value_type, load_value_type
+from ._multi_value import MultiValueTypeInfo
+from ._single_value import (
     AssetEntity,
     AssetFields,
     ClassEntity,
@@ -17,9 +20,6 @@ from ._base import (
     UnknownEntity,
     ViewEntity,
 )
-from ._constants import EntityTypes, Undefined, Unknown
-from ._loaders import load_connection, load_dms_value_type, load_value_type
-from ._multi_value import MultiValueTypeInfo
 from ._types import CdfResourceEntityList, ClassEntityList, ContainerEntityList, URLEntity, ViewEntityList
 from ._wrapped import DMSFilter, HasDataFilter, NodeTypeFilter, RawFilter, WrappedEntity
