+++ cognite/neat/rules/models/entities/_single_value.py
